make auto generation a util function
