+++ offscreen-canvas/tools/templates.yaml
@@ -40,6 +40,7 @@ w3cmain: |
     </script>
 
 w3cworker: |
+    // DO NOT EDIT! This test has been generated by tools/gentest.py.
     importScripts("/resources/testharness.js");
     importScripts("/resources/testharnessreport.js");
     importScripts("/common/canvas-tests.js");
